requirements.txt
fastapi
uvicorn
openai>=1.0.0
pydantic
python-dotenv
streamlit
pandas
